grouped server tests
