+++ integration-tests/server/index.test.js
@@ -1,0 +1,5 @@
+import connect from './connect';
+import power from './power';
+
+connect();
+power();
