+++ integration-tests/server/power.js
@@ -1,0 +1,13 @@
+import createServer from '@arranger/server';
+
+export default () =>
+  test('server should start and stop no demand', () => {
+    let server = createServer();
+    expect(server.status).toBe('off');
+    server.listen(7357, () => {
+      expect(server.status).toBe('on');
+      server.close(() => {
+        expect(server.status).toBe('off');
+      });
+    });
+  });
